@@ -3,17 +3,17 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.14.0",
     "@formspree/react": "^2.2.4",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^12.0.0",
     "@testing-library/user-event": "^13.2.1",
     "@types/jest": "^27.0.1",
     "@types/node": "^16.7.13",
-    "@types/react": "^17.0.20",
-    "@types/react-dom": "^17.0.9",
     "@vercel/analytics": "^1.4.1",
     "react": "^18.3.1",
-    "react-animate-on-scroll": "^2.1.5",
+    "react-awesome-reveal": "^4.3.1",
     "react-dom": "^18.3.1",
     "react-google-recaptcha": "^2.1.0",
     "react-router-dom": "^6.2.1",
@@ -22,7 +22,7 @@
     "react-toastify": "^8.1.0",
     "react-tsparticles": "^1.37.5",
     "styled-components": "^5.3.3",
-    "typescript": "^4.4.2",
+    "typescript": "^5.8.3",
     "validator": "^13.7.0",
     "web-vitals": "^2.1.0"
   },
@@ -51,7 +51,8 @@
     ]
   },
   "devDependencies": {
-    "@types/react-animate-on-scroll": "^2.1.5",
+    "@types/react": "^19.1.6",
+    "@types/react-dom": "^19.1.6",
     "@types/react-google-recaptcha": "^2.1.2",
     "@types/react-router-hash-link": "^2.4.4",
     "@types/styled-components": "^5.1.18",
