@@ -12,107 +12,106 @@ import reactIcon from "../../assets/react-icon.svg";
 import typescriptIcon from "../../assets/typescript-icon.svg";
 import vueIcon from "../../assets/vue-icon.svg";
 import boostrapIcon from "../../assets/bootstrap-icon.svg";
-import ScrollAnimation from "react-animate-on-scroll";
-
+import { Fade } from 'react-awesome-reveal';
 export function About() {
   return (
     <Container id="about">
       <div className="about-text">
-        <ScrollAnimation animateIn="fadeInLeft">
+        <Fade animateIn="fadeInLeft">
           <h2>About me</h2>
-        </ScrollAnimation>
-        <ScrollAnimation animateIn="fadeInLeft" delay={0.1 * 1000}>
+        </Fade>
+        <Fade animateIn="fadeInLeft" delay={0.1 * 1000}>
           <p>
           Hi there! I'm Meet Navapara, a passionate Full Stack Developer with hands-on experience in building scalable web applications using modern JavaScript frameworks like React and Node.js.
              </p>
-        </ScrollAnimation>
-        <ScrollAnimation animateIn="fadeInLeft" delay={0.2 * 1000} style={{ marginTop: "2rem", marginBottom: "2rem" }}>
+        </Fade>
+        <Fade animateIn="fadeInLeft" delay={0.2 * 1000} style={{ marginTop: "2rem", marginBottom: "2rem" }}>
           <p>
           I specialize in crafting responsive front-end interfaces, integrating RESTful APIs, and designing efficient back-end architectures with Express and MongoDB.
           </p>
-        </ScrollAnimation>
-        <ScrollAnimation animateIn="fadeInLeft" delay={0.3 * 1000}>
+        </Fade>
+        <Fade animateIn="fadeInLeft" delay={0.3 * 1000}>
           <p>
           I thrive in Agile environments, collaborate seamlessly with cross-functional teams, and have a strong focus on delivering smooth user experiences and optimized performance.
           </p>
-        </ScrollAnimation>
-        <ScrollAnimation animateIn="fadeInLeft" delay={400}>
+        </Fade>
+        <Fade animateIn="fadeInLeft" delay={400}>
           <div className="education">
             <h3>Education:</h3>
             <h4> Master of Computer Applications (MCA)</h4>
             <p>Jain University, Bangalore | July 2023 - Present</p>
             <p>9.0 CGPA</p>
           </div>
-        </ScrollAnimation>
-        <ScrollAnimation animateIn="fadeInLeft" delay={550}>
+        </Fade>
+        <Fade animateIn="fadeInLeft" delay={550}>
           <div className="experience">
              <h3>Experience:</h3><h4>Software Developer</h4>
             <p>Binary Bird Technologies | February 2022 - April 2025</p>
             <p>Surat, India</p>
           </div>
-        </ScrollAnimation>
+        </Fade>
 
-        <ScrollAnimation animateIn="fadeInLeft" delay={0.4 * 1000}>
+        <Fade animateIn="fadeInLeft" delay={0.4 * 1000}>
           <h3>Here are my main skills:</h3>
-        </ScrollAnimation>
+        </Fade>
         <div className="hard-skills">
          
           <div className="hability">
-            <ScrollAnimation animateIn="fadeInUp" delay={0.12 * 1000}>
+            <Fade animateIn="fadeInUp" delay={0.12 * 1000}>
               <img src={jsIcon} alt="JavaScript" />
-            </ScrollAnimation>
+            </Fade>
           </div>
           <div className="hability">
-            <ScrollAnimation animateIn="fadeInUp" delay={0.13 * 1000}>
+            <Fade animateIn="fadeInUp" delay={0.13 * 1000}>
               <img src={reactIcon} alt="React" />
-            </ScrollAnimation>
+            </Fade>
           </div>
           <div className="hability">
-            <ScrollAnimation animateIn="fadeInUp" delay={0.14 * 1000}>
+            <Fade animateIn="fadeInUp" delay={0.14 * 1000}>
               <img src={express} alt="Express" />
-            </ScrollAnimation>
+            </Fade>
           </div>
           <div className="hability">
-            <ScrollAnimation animateIn="fadeInUp" delay={0.14 * 1000}>
+            <Fade animateIn="fadeInUp" delay={0.14 * 1000}>
               <img src={typescriptIcon} alt="Typescript" />
-            </ScrollAnimation>
+            </Fade>
           </div>
           <div className="hability">
-            <ScrollAnimation animateIn="fadeInUp" delay={0.15 * 1000}>
+            <Fade animateIn="fadeInUp" delay={0.15 * 1000}>
               <img src={vueIcon} alt="Vue" />
-            </ScrollAnimation>
+            </Fade>
           </div>
           <div className="hability">
-            <ScrollAnimation animateIn="fadeInUp" delay={0.15 * 1000}>
+            <Fade animateIn="fadeInUp" delay={0.15 * 1000}>
               <img src={nodejs} alt="node" />
-            </ScrollAnimation>
+            </Fade>
           </div>
           <div className="hability">
-            <ScrollAnimation animateIn="fadeInUp" delay={0.15 * 1000}>
+            <Fade animateIn="fadeInUp" delay={0.15 * 1000}>
               <img src={mongo} alt="Mongodb" />
-            </ScrollAnimation>
+            </Fade>
           </div>
           <div className="hability">
-            <ScrollAnimation animateIn="fadeInUp" delay={0.18 * 1000}>
+            <Fade animateIn="fadeInUp" delay={0.18 * 1000}>
               <img src={htmlIcon} alt="Html" />
-            </ScrollAnimation>
+            </Fade>
           </div>
           <div className="hability">
-            <ScrollAnimation animateIn="fadeInUp" delay={0.19 * 1000}>
+            <Fade animateIn="fadeInUp" delay={0.19 * 1000}>
               <img src={cssIcon} alt="Css" />
-            </ScrollAnimation>
+            </Fade>
           </div>
           <div className="hability">
-            <ScrollAnimation animateIn="fadeInUp" delay={0.20 * 1000}>
+            <Fade animateIn="fadeInUp" delay={0.20 * 1000}>
               <img src={boostrapIcon} alt="bootstrap" />
-            </ScrollAnimation>
+            </Fade>
           </div>
         </div>
       </div>
       <div className="about-image">
-        <ScrollAnimation animateIn="fadeInRight" delay={0.21 * 1000}>
+        <Fade animateIn="fadeInRight" delay={0.21 * 1000}>
           <img src={Meet} alt="Meet Navapara" />
-        </ScrollAnimation>
+        </Fade>
       </div>
     </Container>
   )
