@@ -1,6 +1,6 @@
 import { BrowserRouter } from "react-router-dom"
 import { Container } from "./styles"
-import ScrollAnimation from "react-animate-on-scroll"
+import { Fade } from 'react-awesome-reveal';
 import Illustration from "../../assets/illustration.svg"
 import { NavHashLink } from "react-router-hash-link"
 import linkedin from '../../assets/linkedin.svg'
@@ -12,24 +12,24 @@ export function Hero() {
   return (
     <Container id="home">
       <div className="hero-text">
-        <ScrollAnimation animateIn="fadeInUp">
+        <Fade animateIn="fadeInUp">
           <p>Hello <img src={Hello} alt="Hello" width="20px"/>, I'm</p>
-        </ScrollAnimation>
-        <ScrollAnimation animateIn="fadeInUp" delay={0.2 * 1000}>
+        </Fade>
+        <Fade animateIn="fadeInUp" delay={0.2 * 1000}>
           <h1>Meet Navapara</h1>
-        </ScrollAnimation>
-        <ScrollAnimation animateIn="fadeInUp" delay={0.4 * 1000}>
+        </Fade>
+        <Fade animateIn="fadeInUp" delay={0.4 * 1000}>
           <h3>Full stack developer</h3>
-        </ScrollAnimation>
-        <ScrollAnimation animateIn="fadeInUp" delay={0.6 * 1000}>
+        </Fade>
+        <Fade animateIn="fadeInUp" delay={0.6 * 1000}>
           <p className="small-resume">3 Year Experience</p>
-        </ScrollAnimation>
-        <ScrollAnimation animateIn="fadeInUp" delay={0.8 * 1000}>
+        </Fade>
+        <Fade animateIn="fadeInUp" delay={0.8 * 1000}>
           <BrowserRouter>
             <NavHashLink smooth to="#contact" className="button">Contact</NavHashLink>
           </BrowserRouter>
-        </ScrollAnimation>
-        <ScrollAnimation animateIn="fadeInUp" delay={1 * 1000}>
+        </Fade>
+        <Fade animateIn="fadeInUp" delay={1 * 1000}>
       <div className="social-media"><a
         href="https://www.linkedin.com/in/codevinayak"
         target="_blank"
@@ -58,12 +58,12 @@ export function Hero() {
         >
           <img src={telegram} alt="telegram" />
         </a></div>
-        </ScrollAnimation>
+        </Fade>
       </div>
       <div className="hero-image">
-        <ScrollAnimation animateIn="fadeInRight" delay={1 * 1000}>
+        <Fade animateIn="fadeInRight" delay={1 * 1000}>
           <img src={Illustration} alt="Ilustração" />
-        </ScrollAnimation>
+        </Fade>
       </div>
     </Container>
   )
