@@ -2,8 +2,7 @@ import { Container } from "./styles";
 import githubIcon from "../../assets/github.svg"
 // import DownloadApp from '../../assets/download.webp'
 import externalLink from "../../assets/external-link.svg"
-import ScrollAnimation from "react-animate-on-scroll";
-
+import { Fade } from 'react-awesome-reveal';
 
 export function Project() {
   return (
@@ -13,7 +12,7 @@ export function Project() {
 
   
 
-        <ScrollAnimation animateIn="flipInX">
+        <Fade animateIn="flipInX">
           <div className="project">
             <header>
               <svg width="50" xmlns="http://www.w3.org/2000/svg" role="img" viewBox="0 0 24 24" fill="none" stroke="#23ce6b " stroke-width="1" stroke-linecap="round" stroke-linejoin="round" ><title>Folder</title><path d="M22 19a2 2 0 0 1-2 2H4a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h5l2 3h9a2 2 0 0 1 2 2z"></path></svg>
@@ -41,9 +40,9 @@ export function Project() {
               </ul>
             </footer>
           </div>
-        </ScrollAnimation>
+        </Fade>
 
-        <ScrollAnimation animateIn="flipInX">
+        <Fade animateIn="flipInX">
           <div className="project">
             <header>
               <svg width="50" xmlns="http://www.w3.org/2000/svg" role="img" viewBox="0 0 24 24" fill="none" stroke="#23ce6b" strokeWidth="1" strokeLinecap="round" strokeLinejoin="round">
@@ -74,9 +73,9 @@ export function Project() {
               </ul>
             </footer>
           </div>
-        </ScrollAnimation>
+        </Fade>
 
-        <ScrollAnimation animateIn="flipInX">
+        <Fade animateIn="flipInX">
           <div className="project">
             <header>
               <svg width="50" xmlns="http://www.w3.org/2000/svg" role="img" viewBox="0 0 24 24" fill="none" stroke="#23ce6b" strokeWidth="1" strokeLinecap="round" strokeLinejoin="round">
@@ -109,10 +108,10 @@ export function Project() {
               </ul>
             </footer>
           </div>
-        </ScrollAnimation>
+        </Fade>
 
         
-        {/* <ScrollAnimation animateIn="flipInX">
+        {/* <Fade animateIn="flipInX">
           <div className="project">
             <header>
               <svg width="50" xmlns="http://www.w3.org/2000/svg" role="img" viewBox="0 0 24 24" fill="none" stroke="#23ce6b " stroke-width="1" stroke-linecap="round" stroke-linejoin="round" ><title>Folder</title><path d="M22 19a2 2 0 0 1-2 2H4a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h5l2 3h9a2 2 0 0 1 2 2z"></path></svg>
@@ -137,9 +136,9 @@ export function Project() {
               </ul>
             </footer>
           </div>
-        </ScrollAnimation> */}
+        </Fade> */}
 
-        {/* <ScrollAnimation animateIn="flipInX">
+        {/* <Fade animateIn="flipInX">
           <div className="project">
             <header>
               <svg width="50" xmlns="http://www.w3.org/2000/svg" role="img" viewBox="0 0 24 24" fill="none" stroke="#23ce6b " stroke-width="1" stroke-linecap="round" stroke-linejoin="round" ><title>Folder</title><path d="M22 19a2 2 0 0 1-2 2H4a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h5l2 3h9a2 2 0 0 1 2 2z"></path></svg>
@@ -162,9 +161,9 @@ export function Project() {
               </ul>
             </footer>
           </div>
-        </ScrollAnimation> */}
+        </Fade> */}
 
-        {/* <ScrollAnimation animateIn="flipInX">
+        {/* <Fade animateIn="flipInX">
           <div className="project">
             <header>
               <svg width="50" xmlns="http://www.w3.org/2000/svg" role="img" viewBox="0 0 24 24" fill="none" stroke="#23ce6b " stroke-width="1" stroke-linecap="round" stroke-linejoin="round" ><title>Folder</title><path d="M22 19a2 2 0 0 1-2 2H4a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h5l2 3h9a2 2 0 0 1 2 2z"></path></svg>
@@ -186,9 +185,9 @@ export function Project() {
               </ul>
             </footer>
           </div>
-        </ScrollAnimation> */}
+        </Fade> */}
 
-        {/* <ScrollAnimation animateIn="flipInX">
+        {/* <Fade animateIn="flipInX">
           <div className="project">
             <header>
               <svg width="50" xmlns="http://www.w3.org/2000/svg" role="img" viewBox="0 0 24 24" fill="none" stroke="#23ce6b " stroke-width="1" stroke-linecap="round" stroke-linejoin="round" ><title>Folder</title><path d="M22 19a2 2 0 0 1-2 2H4a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h5l2 3h9a2 2 0 0 1 2 2z"></path></svg>
@@ -210,10 +209,10 @@ export function Project() {
               </ul>
             </footer>
           </div>
-        </ScrollAnimation> */}
+        </Fade> */}
 
         {/* 
-        <ScrollAnimation animateIn="flipInX">
+        <Fade animateIn="flipInX">
           <div className="project">
             <header>
               <svg width="50" xmlns="http://www.w3.org/2000/svg" role="img" viewBox="0 0 24 24" fill="none" stroke="#23ce6b " stroke-width="1" stroke-linecap="round" stroke-linejoin="round" ><title>Folder</title><path d="M22 19a2 2 0 0 1-2 2H4a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h5l2 3h9a2 2 0 0 1 2 2z"></path></svg>
@@ -235,9 +234,9 @@ export function Project() {
               </ul>
             </footer>
           </div>
-        </ScrollAnimation> */}
+        </Fade> */}
 
-        {/* <ScrollAnimation animateIn="flipInX">
+        {/* <Fade animateIn="flipInX">
           <div className="project">
             <header>
               <svg width="50" xmlns="http://www.w3.org/2000/svg" role="img" viewBox="0 0 24 24" fill="none" stroke="#23ce6b " stroke-width="1" stroke-linecap="round" stroke-linejoin="round" ><title>Folder</title><path d="M22 19a2 2 0 0 1-2 2H4a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h5l2 3h9a2 2 0 0 1 2 2z"></path></svg>
@@ -259,9 +258,9 @@ export function Project() {
               </ul>
             </footer>
           </div>
-        </ScrollAnimation> */}
+        </Fade> */}
 
-        {/* <ScrollAnimation animateIn="flipInX">
+        {/* <Fade animateIn="flipInX">
           <div className="project">
             <header>
               <svg width="50" xmlns="http://www.w3.org/2000/svg" role="img" viewBox="0 0 24 24" fill="none" stroke="#23ce6b " stroke-width="1" stroke-linecap="round" stroke-linejoin="round" ><title>Folder</title><path d="M22 19a2 2 0 0 1-2 2H4a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h5l2 3h9a2 2 0 0 1 2 2z"></path></svg>
@@ -284,7 +283,7 @@ export function Project() {
               </ul>
             </footer>
           </div>
-        </ScrollAnimation> */}
+        </Fade> */}
 
       </div>
     </Container>
